Added Example File
Added an example file to allow testing in browser.  Usage is described in
README.

Added a title attribute to fields -- Needs testing

Changed the default template to use fieldsets and legends for forms.
Added a submit button. (What, you mean you want to submit the form? )

Changed the field template to display labels.