More String Validations
Added String validations for 'match', 'length', and 'min_length'

Added a format parameter to fields.  At the moment it is only used to
change text fields to password fields when it's value is set to hidden.
This is bordering on requiring the model to know about how the
interface will display a field, but it solves the problem for now.

Form#add_field now always returnd self

All Test passed